core: splitting all files into 3 main packages
activities package
adapters package
models package
