+++ Kidzenia/app/src/main/AndroidManifest.xml
@@ -10,15 +10,19 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Kidzenia">
         <activity
-            android:name=".VideosActivity"
+            android:name=".activities.VideoPlayerActivity"
             android:exported="false" />
-        <activity android:name=".DisplayDataActivity" />
-        <activity android:name=".MainActivity" />
         <activity
-            android:name=".SplashActivity"
+            android:name=".activities.VideosActivity"
+            android:exported="false" />
+        <activity android:name=".activities.DisplayDataActivity" />
+        <activity android:name=".activities.MainActivity" />
+        <activity
+            android:name=".activities.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
