+++ Kidzenia/app/src/main/java/com/example/android/kidzenia/activities/VideoPlayerActivity.java
@@ -1,0 +1,16 @@
+package com.example.android.kidzenia.activities;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.android.kidzenia.R;
+
+public class VideoPlayerActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_video_player);
+    }
+}
