+++ Kidzenia/app/src/main/res/layout/activity_display_data.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/inside_class"
-    tools:context=".DisplayDataActivity">
+    tools:context=".activities.DisplayDataActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_view"
